win32 u_int cleanup, from mike@datanerds.net

+++ src/rand.c
@@ -6,7 +6,7 @@
  * Copyright (c) 2000 Dug Song <dugsong@monkey.org>
  * Copyright (c) 1996 David Mazieres <dm@lcs.mit.edu>
  *
- * $Id: rand.c,v 1.10 2003/03/06 04:32:25 dugsong Exp $
+ * $Id: rand.c,v 1.11 2003/03/21 04:03:05 dugsong Exp $
  */
 
 #include "config.h"
@@ -17,11 +17,11 @@
 #else
 #include <sys/types.h>
 #include <sys/time.h>
+#include <unistd.h>
 #endif
 #include <fcntl.h>
 #include <stdlib.h>
 #include <string.h>
-#include <unistd.h>
 
 #include "dnet.h"
 
@@ -111,7 +111,7 @@ int
 rand_get(rand_t *r, void *buf, size_t len)
 {
 	u_char *p;
-	int i;
+	u_int i;
 
 	for (p = buf, i = 0; i < len; i++) {
 		p[i] = rand_getbyte(r);
@@ -167,12 +167,12 @@ int
 rand_shuffle(rand_t *r, void *base, size_t nmemb, size_t size)
 {
 	u_char *save, *src, *dst, *start = (u_char *)base;
-	int i, j;
+	u_int i, j;
 
 	if (nmemb < 2)
 		return (0);
 	
-	if (r->tmplen < size) {
+	if ((u_int)r->tmplen < size) {
 		if (r->tmp == NULL) {
 			if ((save = malloc(size)) == NULL)
 				return (-1);
